refactor(book): unify rating field type to Double

+++ src/main/java/com/maximebaron/springbook/book/command/UpdateBookCommand.java
@@ -24,5 +24,5 @@ public record UpdateBookCommand(
         LocalDate publishedAt,
 
         @Min(0) @Max(5)
-        Short rating
+        Double rating
 ) {}
