+++ src/main/java/com/maximebaron/springbook/book/dto/CreateBookRequest.java
@@ -23,5 +23,5 @@ public record CreateBookRequest(
 
         @NotNull
         @Min(0) @Max(5)
-        Short rating
+        Double rating
 ) {}
